systemvmtemplate: new template for 4.15.1
